1. Removed preview/media.js 2. Renamed core/popcorn.js to core/popcorn-wrapper.js and updated references.
