+++ src/core/popcorn-wrapper.js
